Move NEWS from linux-2.6 to linux-latest-2.6
Every kernel ABI bump introduces new binary packages, which are
normally installed as dependencies of an upgraded linux-image package
built from linux-latest-2.6.

Since apt-listchanges only shows NEWS for upgraded packages, not newly
installed packages, we should put it in the linux-latest-2.6 packages.
Backdate the description of the libata transition to the release of
linux-latest-2.6 which depends on it.

This also means that NEWS won't be shown for linux-libc-dev upgrades
which it doesn't affect. (Closes: #586401)

svn path=/dists/sid/linux-latest-2.6/; revision=15899