linux-image-{686-pae,amd64}: Add backdated NEWS for introduction of xz compression
This problem (#727736) affects domU kernels in old versions of Xen,
and these are the only two flavours that run in PV mode.

There is no need to tell users already running testing/unstable about
this, as it either doesn't affect them or they've already hit the bug.
I've backdated the NEWS entry to the first version in unstable that
depended on an xz-compressed kernel, so it should only appear for
users upgrading from wheezy.

svn path=/dists/sid/linux-latest/; revision=21728